Add files via upload
Starting version of the script, 10/4.
main.py contains the construction of the dataset, the running of the machine, the test phase and the I/O interface.
RBM_matrix instead contains the actual implementation. "Matrix" refers to the fact that, w.r.t. the alpha version of the script, learning is optimized through matrix-matrix products.