Towards IAS implementation.
Defined a free energy function for the two types of RBMs, that can be used to get the values of the non-normalized marginalized distribution p_* (in the notation of Salakhutidnov). 
Defined the energy function for the ReLU_RBM, following the eq. (1) of Monasson's article.  
Optimized monitorOverfitting using matrix-matrix multiplications for the BaseRBM.
Minor change in _init__