AIS implementation (alpha version)
Defined the AIS function following Salakhutdinov's article. 
Modified the free energy function in order to consider the varying temperatures of the two considered RBMs.
Current numerical issue that returns a null value when computing the denominator of the importance weights.
The thresholds do not appear to be crucial for this behaviour,  even though it should be related to the $\theta < \infty$ constraint according to the article.