Last version of the ReLU_RBM
Modified slightly the CD function to reduce noise sampling and reconduct to Hinton's CD formula.
Even if it is more rigorous, it reduces marginally the performance. 
Modified the I/O interface (in particular the output to be saved in dataset.csv) and introduced the computation of the mean squared activity of the two types of hidden units, i.e. strongly activated ones and the others.