Add files via upload
Fixed the problem concerning the mini-batches, namely now fit receives the whole training set and splits it according to the number of minibatches. 
Added a condition in the generation of the dataset to make possible the use of the MNIST dataset. 
Modified AIS to make the computation of the weights importance more stable (but it can be still improved and the problem of the positive likelihood is still occuring).
Moved to PCD the update of the chains before every gradient computation (if the chains are already defined).
Removed __phi in BaseRBM in favour of scipy.special.expit().
Defined a control over the growing alpha to avoid instability, i.e. alpha <= 1 always. 
Modified gibbsSampling (both versions) according to section 3 of Hinton's guide, that is the machine now daydreams using always the hidden states in each Gibbs step together with the probabilities for the visible layer, to reduce noise sampling.